--- sd/h2/src/main/java/ru/bauman/tigerbank/analytic/AnalyticsServiceInterface.java
@@ -1,12 +1,0 @@
-package ru.bauman.tigerbank.analytic;
-
-import ru.bauman.tigerbank.category.dto.CategoryDto;
-import ru.bauman.tigerbank.operation.entity.OperationTypeEnum;
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.Map;
-
-public interface AnalyticsServiceInterface {
-	BigDecimal getIncomeExpenseDifference(Long accountId, LocalDateTime from, LocalDateTime to);
-	Map<CategoryDto, BigDecimal> groupByCategory(Long accountId, LocalDateTime from, LocalDateTime to, OperationTypeEnum typeEnum);
-}
