--- sd/h2/src/main/java/ru/bauman/tigerbank/category/service/CategoryServiceInterface.java
@@ -1,12 +1,0 @@
-package ru.bauman.tigerbank.category.service;
-
-import ru.bauman.tigerbank.category.dto.CategoryDto;
-import java.util.List;
-
-public interface CategoryServiceInterface {
-	CategoryDto create(CategoryDto dto);
-	CategoryDto update(Long id, CategoryDto dto);
-	void delete(Long id);
-	CategoryDto findById(Long id);
-	List<CategoryDto> findAll();
-}
