+++ sd/h2/src/main/java/ru/bauman/tigerbank/category/service/entity/CategoryEntityService.java
@@ -1,9 +1,9 @@
-package ru.bauman.tigerbank.category.service;
+package ru.bauman.tigerbank.category.service.entity;
 
 import ru.bauman.tigerbank.category.entity.Category;
 import java.util.List;
 
-public interface CategoryEntityServiceInterface {
+public interface CategoryEntityService {
 	Category save(Category category);
 	void deleteById(Long id);
 	Category getById(Long id);
