+++ sd/h2/src/main/java/ru/bauman/tigerbank/category/service/entity/CategoryTypeEntityService.java
@@ -1,9 +1,9 @@
-package ru.bauman.tigerbank.category.service;
+package ru.bauman.tigerbank.category.service.entity;
 
 import ru.bauman.tigerbank.category.entity.CategoryType;
 import ru.bauman.tigerbank.category.entity.CategoryTypeEnum;
 
-public interface CategoryTypeEntityServiceInterface {
+public interface CategoryTypeEntityService {
 	CategoryType getById(Long id);
 	CategoryType getByName(CategoryTypeEnum name);
 }
